APIs/presalytics.io/ooxml/0.1.0/python-aiohttp-client/tests/test_slides_graphic_types_controller.py
# coding: utf-8

import pytest
import json
from aiohttp import web

from openapi_server.models.problem_details import ProblemDetails
from openapi_server.models.slide_graphic_types import SlideGraphicTypes


pytestmark = pytest.mark.asyncio

async def test_slides_graphictypes_get(client):
    """Test case for slides_graphictypes_get

    GraphicTypes: List All Possible Types
    """
    headers = { 
        'Accept': 'application/json',
    }
    response = await client.request(
        method='GET',
        path='/ooxml-automation/Slides/GraphicTypes',
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_slides_graphictypes_get_id(client):
    """Test case for slides_graphictypes_get_id

    GraphicTypes: Get by Id
    """
    headers = { 
        'Accept': 'application/json',
    }
    response = await client.request(
        method='GET',
        path='/ooxml-automation/Slides/GraphicTypes/{id}'.format(id='id_example'),
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_slides_graphictypes_typeid_get_type_id(client):
    """Test case for slides_graphictypes_typeid_get_type_id

    GraphicTypes: Get By Type Id
    """
    headers = { 
        'Accept': 'application/json',
    }
    response = await client.request(
        method='GET',
        path='/ooxml-automation/Slides/GraphicTypes/TypeId/{type_id}'.format(type_id=56),
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')

